+++ tests/tests.nim
@@ -10,7 +10,8 @@ import units/[
   tio,
   tir,
   tnotes,
-  ttext
+  ttext,
+  ttracking
 ]
 
 import units/private/[